requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
sentence-transformers>=2.2
transformers>=4.21.0
torch>=1.12.0
rapidfuzz>=3.5
python-dotenv>=1.0
matplotlib>=3.7
seaborn>=0.12
faiss-cpu>=1.7.0
pyarrow>=10.0.0
hdbscan>=0.8.29
openai>=1.0.0
